refactor(kernel): flatten cache/ — move index_cache into core/ per §4
Third move in the §4 alignment series (after cas_* → core/cas/ in
`dc9ad261d` and rpc_transport → core/ in `95941084d`). The
crate-root `cache/` directory contained a single file
(`index_cache.rs`) and was inconsistent with `core/permission_cache.rs`,
which already sits flat inside `core/`.

* `rust/kernel/src/cache/index_cache.rs` → `core/index_cache.rs`.
* Delete the now-empty `cache/mod.rs`.
* Drop `pub mod cache;` from the crate root.
* Add `pub mod index_cache;` to `core/mod.rs`.
* Rewrite the three internal callsites
  (`kernel/mod.rs` × 2, `kernel/io.rs` × 1) from
  `crate::cache::index_cache::*` → `crate::core::index_cache::*`.

No external Rust caller referenced `kernel::cache::*`; no
re-export shim needed. Same seven-criteria pass as the previous
two moves: simpler contract (every cache primitive now in
`core/`), architecturally consistent with §4, SSOT preserved.

`cargo check --workspace --tests` clean.
