--- rust/kernel/src/cache/mod.rs
@@ -1,1 +1,0 @@
-pub mod index_cache;
