+++ rust/kernel/src/core/index_cache.rs
